feat: parallel batch [T20260509-1]
Tasks:
- T20260509-1: Modularize crates/orbit-cli/src/command/web/api.rs (3,376 lines)
